requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1

SQLAlchemy>=2.0.30
pillow>=10.3.0